PortfolioBackendDjango/requirements.txt
urllib3<2
django
djangorestframework
requests
python-decouple
django-cors-headers